Fixed rebase and moved js/html into jsx
