--- webofneeds/won-owner-webapp/src/main/webapp/app/components/connections/connections.html
@@ -1,40 +1,0 @@
-<won-modal-dialog ng-if="self.showModalDialog"></won-modal-dialog>
-<div class="won-modal-atomview" ng-if="self.showAtomOverlay">
-    <won-post-info include-header="true" atom-uri="self.viewAtomUri"></won-post-info>
-</div>
-<div class="won-modal-connectionview" ng-if="self.showConnectionOverlay">
-    <won-post-messages connection-uri="self.viewConnUri"></won-post-messages>
-</div>
-<header>
-    <won-topnav></won-topnav>
-</header>
-<won-toasts></won-toasts>
-<won-slide-in ng-if="self.showSlideIns"></won-slide-in>
-<aside class="overview__left" ng-class="{'hide-in-responsive': self.hideListSideInResponsive}">
-    <won-connections-overview on-selected-connection="::self.selectConnection(connectionUri)"
-                              open="self.open">
-    </won-connections-overview>
-</aside>
-<!-- RIGHT SIDE -->
-<main class="overview__right" ng-if="self.showWelcomeSide" ng-class="{'hide-in-responsive' : self.hideWelcomeSideInResponsive}">
-    <div class="overview__right__welcome">
-        <div class="overview__right__welcome__text" ng-include="self.welcomeTemplatePath"></div>
-        <won-usecase-picker></won-usecase-picker>
-        <div class="overview__right__welcome__howto">
-            <div  class="overview__right__welcome__howto__label">{{ self.appTitle }} connects people's atoms to their solutions</div>
-            <a class="overview__right__welcome__howto__button won-button--filled red" href="{{ ::self.absHRef(self.$state, 'about', {'aboutSection': 'aboutHowTo'}) }}">
-                <span>How {{ self.appTitle }} works?</span>
-            </a>
-        </div>
-    </div>
-</main>
-<main class="overview__right" ng-if="self.showContentSide">
-    <won-usecase-picker ng-if="self.showUseCasePicker"></won-usecase-picker>
-    <won-usecase-group ng-if="self.showUseCaseGroups"></won-usecase-group>
-    <won-create-post ng-if="self.showCreatePost"></won-create-post>
-    <won-create-search ng-if="self.showCreateSearch"></won-create-search>
-    <won-post-messages ng-if="self.showPostMessages"></won-post-messages>
-    <won-group-post-messages ng-if="self.showGroupPostMessages"></won-group-post-messages>
-</main>
-<won-footer ng-class="{'hide-in-responsive': self.hideFooterInResponsive }"> <!-- Connection view does not show the footer in responsive mode as there should not be two scrollable areas imho-->
-</won-footer>
