--- webofneeds/won-owner-webapp/src/main/webapp/app/components/inventory/inventory.html
@@ -1,37 +1,0 @@
-<won-modal-dialog ng-if="self.showModalDialog"></won-modal-dialog>
-<div class="won-modal-atomview" ng-if="self.showAtomOverlay">
-    <won-post-info include-header="true" atom-uri="self.viewAtomUri"></won-post-info>
-</div>
-<div class="won-modal-connectionview" ng-if="self.showConnectionOverlay">
-    <won-post-messages connection-uri="self.viewConnUri"></won-post-messages>
-</div>
-<header>
-    <won-topnav></won-topnav>
-</header>
-<won-toasts></won-toasts>
-<won-slide-in ng-if="self.showSlideIns"></won-slide-in>
-<main class="ownerinventory">
-    <div class="ownerinventory__header">
-        <div class="ownerinventory__header__title">
-            Open <span class="ownerinventory__header__title__count" ng-if="self.hasOwnedActiveAtomUris">({{ self.sortedOwnedActiveAtomUriSize }})</span>
-        </div>
-    </div>
-    <div class="ownerinventory__content" ng-if="self.hasOwnedActiveAtomUris">
-        <won-atom-card class="ownerinventory__content__atom" atom-uri="atomUri" current-location="self.currentLocation" ng-repeat="atomUri in self.sortedOwnedActiveAtomUriArray track by atomUri"></won-atom-card>
-    </div>
-    <div class="ownerinventory__noresults" ng-if="!self.hasOwnedActiveAtomUris">
-        <span class="ownerinventory__noresults__label">Nothing to display</span>
-    </div>
-    <div class="ownerinventory__header">
-        <div class="ownerinventory__header__title">
-            Closed <span class="ownerinventory__header__title__count" ng-if="self.hasOwnedInactiveAtomUris">({{ self.sortedOwnedInactiveAtomUriSize }})</span>
-        </div>
-    </div>
-    <div class="ownerinventory__content" ng-if="self.hasOwnedInactiveAtomUris">
-        <won-atom-card class="ownerinventory__content__atom" atom-uri="atomUri" current-location="self.currentLocation" ng-repeat="atomUri in self.sortedOwnedInactiveAtomUriArray track by atomUri"></won-atom-card>
-    </div>
-    <div class="ownerinventory__noresults" ng-if="!self.hasOwnedInactiveAtomUris">
-        <span class="ownerinventory__noresults__label">Nothing to display</span>
-    </div>
-</main>
-<won-footer></won-footer>
